+++ src/main/kotlin/net/pwall/json/JSONConfig.kt
@@ -25,6 +25,7 @@
 
 package net.pwall.json
 
+import net.pwall.json.annotation.JSONAllowExtra
 import kotlin.reflect.KClass
 import kotlin.reflect.KProperty
 import kotlin.reflect.KType
@@ -73,6 +74,9 @@ class JSONConfig {
     /** Switch to control whether null fields in objects are output as "null": `true` -> yes, `false` -> no */
     var includeNulls = defaultIncludeNulls
 
+    /** Switch to control whether extra fields are allowed on deserialization: `true` -> yes, `false` -> no */
+    var allowExtra = defaultAllowExtra
+
     private val fromJSONMap: MutableMap<KType, FromJSONMapping> = LinkedHashMap()
 
     private val toJSONMap: MutableMap<KType, ToJSONMapping> = LinkedHashMap()
@@ -85,6 +89,8 @@ class JSONConfig {
     private val includeIfNullAnnotations: MutableList<KClass<*>> = arrayListOf(JSONIncludeIfNull::class)
 
     private val includeAllPropertiesAnnotations: MutableList<KClass<*>> = arrayListOf(JSONIncludeAllProperties::class)
+
+    private val allowExtraPropertiesAnnotations: MutableList<KClass<*>> = arrayListOf(JSONAllowExtra::class)
 
     /**
      * Find a `fromJSON` mapping function that will create the specified [KType], or the closest subtype of it.
@@ -337,7 +343,7 @@ class JSONConfig {
 
     /**
      * Add an annotation specification to the list of annotations that specify that all properties in a class are to be
-     * included when serializing or deserializing even if `null`.
+     * included when serializing even if `null`.
      *
      * @param   ignoreAllPropertiesAnnotationClass  the annotation class
      * @param   T                                   the annotation class
@@ -347,14 +353,34 @@ class JSONConfig {
     }
 
     /**
-     * Test whether a property has an annotation to indicate that it is to be included when serializing or deserializing
-     * even if `null`.
+     * Test whether a property has an annotation to indicate that it is to be included when serializing even if `null`.
      *
      * @param   annotations the [Annotation]s (from the class)
      * @return              `true` if an "include all properties" annotation appears in the supplied list
      */
     fun hasIncludeAllPropertiesAnnotation(annotations: List<Annotation>?) =
             hasBooleanAnnotation(includeAllPropertiesAnnotations, annotations)
+
+    /**
+     * Add an annotation specification to the list of annotations that specify that extra properties in a class are to
+     * be ignored when deserializing.
+     *
+     * @param   allowExtraPropertiesAnnotationClass the annotation class
+     * @param   T                                   the annotation class
+     */
+    fun <T: Annotation> addAllowExtraPropertiesAnnotation(allowExtraPropertiesAnnotationClass: KClass<T>) {
+        allowExtraPropertiesAnnotations.add(allowExtraPropertiesAnnotationClass)
+    }
+
+    /**
+     * Test whether a property has an annotation to indicate that extra properties in a class are to be ignored when
+     * deserializing.
+     *
+     * @param   annotations the [Annotation]s (from the class)
+     * @return              `true` if an "allow extra properties" annotation appears in the supplied list
+     */
+    fun hasAllowExtraPropertiesAnnotation(annotations: List<Annotation>?) =
+            hasBooleanAnnotation(allowExtraPropertiesAnnotations, annotations)
 
     /**
      * Test whether a property has a boolean annotation matching the specified list.
@@ -385,12 +411,14 @@ class JSONConfig {
         bigIntegerString = config.bigIntegerString
         bigDecimalString = config.bigDecimalString
         includeNulls = config.includeNulls
+        allowExtra = config.allowExtra
         fromJSONMap.putAll(config.fromJSONMap)
         toJSONMap.putAll(config.toJSONMap)
         nameAnnotations.addAll(config.nameAnnotations)
         ignoreAnnotations.addAll(config.ignoreAnnotations)
         includeIfNullAnnotations.addAll(config.includeIfNullAnnotations)
         includeAllPropertiesAnnotations.addAll(config.includeAllPropertiesAnnotations)
+        allowExtraPropertiesAnnotations.addAll(config.allowExtraPropertiesAnnotations)
     }
 
     /**
@@ -416,6 +444,8 @@ class JSONConfig {
         const val defaultBigDecimalString = false
 
         const val defaultIncludeNulls = false
+
+        const val defaultAllowExtra = false
 
         val defaultCharset = Charsets.UTF_8