+++ src/test/kotlin/net/pwall/json/JSONConfigTest.kt
@@ -195,6 +195,7 @@ class JSONConfigTest {
             bigIntegerString = true
             bigDecimalString = true
             includeNulls = true
+            allowExtra = true
         }
         val config2 = JSONConfig().apply {
             combineAll(config)
@@ -202,6 +203,7 @@ class JSONConfigTest {
         expect(true) { config2.bigIntegerString }
         expect(true) { config2.bigDecimalString }
         expect(true) { config2.includeNulls }
+        expect(true) { config2.allowExtra }
     }
 
     @Test fun `toJSON mapping of nullable type should be selected correctly`() {